// sampa/src/common/enums/action.enum.ts
export enum ActionEnum {
  // Assessment Request Process Actions
  InitiatedSave = 'initiated_save',
  InitiatedRegister = 'initiated_register',
  DraftRegister = 'draft_register',
  DraftModify = 'draft_modify',
  SubmittedApprove = 'submitted_approve',
  SubmittedDecline = 'submitted_decline',
  SubmittedModify = 'submitted_modify',
  ApprovedAccept = 'approved_accept',
  ApprovedDisapprove = 'approved_disapprove',
  ApprovedModify = 'approved_modify',
  AcceptedArrangeTeam = 'accepted_arrangeTeam',
  AcceptedReconsider = 'accepted_reconsider',
  AwaitingSpecsProvideSpecs = 'awaitingSpecs_provideSpecs',
  PreEvaluationAccept = 'preEvaluation_accept',
  PreEvaluationReject = 'preEvaluation_reject',
  OnboardingFinalizeSpecs = 'onboarding_finalizeSpecs',
  StatusReportedNeedChanges = 'statusReported_needChanges',
  UnderAnalysisWrapUp = 'underAnalysis_wrapUp',
  StatusReportedFinalize = 'statusReported_finalize',
  ClosedReopen = 'closed_reopen',

  // Assessment Layer Process Actions
  ApprovedAssignSAM = 'approved_assignSAM',
  SAMAssignedAssignSAM = 'SAM_assigned_assignSAM',
  SAMAssignedAddSupervise = 'SAM_assigned_add_supervise',
  SupervisedAssignOwnTeamAuditors = 'supervised_assignOwnTeamAuditors',
  SupervisedAssignAllTeamsAuditors = 'supervised_assignAllTeamsAuditors',
  SupervisedAssignSupervisor = 'supervised_assignSupervisor',
  GroupReadyAssignAuditor = 'groupReady_assignAuditor',
  GroupReadyReferApplicant = 'groupReady_referApplicant',
  PendingLayerSpecsSubmit = 'pendingLayerSpecs_submit',
  PendingLayerTestcasesSubmit = 'pendingLayerTestcases_submit',
  LayerSpecPreEvaluationNeedModifications = 'layerSpecPreEvaluation_needModifications',
  LayerSpecPreEvaluationAccept = 'layerSpecPreEvaluation_accept',
  LayerSpecOnboardingAccept = 'layerSpecOnboarding_accept',
  LayerAssessmentCompletedAccept = 'layerAssessmentCompleted_accept',
  LayerAssessmentReviewAcceptBySAM = 'layerAssessmentReview_accept_by_sam',
  LayerAssessmentReviewNeedModificationsBySAM = 'layerAssessmentReview_needModifications_by_sam',
  LayerAssessmentCompletedFeedback = 'layerAssessmentCompleted_feedback',
  LayerAssessmentReviewAccept = 'layerAssessmentReview_accept',
  LayerAssessmentReviewNeedModifications = 'layerAssessmentReview_needModifications',
  LayerAssessmentReviewFinish = 'layerAssessmentReview_finish',
  LayerReportIssuedRefer = 'layerReportIssued_refer',
  LayerReportIssuedRemark = 'layerReportIssued_remark',
  RemediateLayerVulnerabilitiesDecline = 'remediateLayerVulnerabilities_decline',
  RemediateLayerVulnerabilitiesFinalize = 'remediateLayerVulnerabilities_finalize',
  ReviewLayerRemediatesApprove = 'reviewLayerRemediates_approve',
  ReviewLayerRemediatesReject = 'reviewLayerRemediates_reject',
  LayerReEvaluationRequestedAccept = 'layerReEvaluationRequested_accept',
  LayerReEvaluationRequestedReject = 'layerReEvaluationRequested_reject',
  LayerAssessmentFulfilledReinstate = 'layerAssessmentFulfilled_reinstate',

  // Others
  Read = 'read',
  ReadForTeam = 'read for team',
  ReadMine = 'read mine',
  CreateTestCaseComment = 'createTestCaseComment',
  GetTestCaseComment = 'getTestCaseComment',
  DeleteTestCaseComment = 'deleteTestCaseComment',
  CreateRemediateComment = 'createRemediateComment',
  GetRemediateComment = 'getRemediateComment',
  DeleteRemediateComment = 'deleteRemediateComment',
  CreateRequestComment = 'createRequestComment',
  GetRequestComment = 'getRequestComment',
  CreateSpecComment = 'createSpecComment',
  GetSpecComment = 'getSpecComment',
  DeleteSpecComment = 'deleteSpecComment',
  CreateLayerComment = 'createLayerComment',
  GetLayerComment = 'getLayerComment',
  DeleteLayerComment = 'deleteLayerComment',
  ApplicantManagerTeamup = 'applicant manager teamup',
  ReadVulnerabilityCount = 'Read Vulnerability Count',
  ReadActionLog = 'Read Action Log',
  ReadSpecs = 'read specs',
  ReadRemediate = 'read_remediate',
  ReadTestCaseMine = 'read test case mine',
  ReadTestCaseAfterFirstIteration = 'Read Test Case After First Iteration',
  ReadTestCase = 'read test case',
  ReadLayers = 'read layers',
  ReadWhereIAmManager = 'read where I am manager',
  ReadLayersForTeam = 'read layers for team',
}